+++ CosmoQuestBot-master/commands/example.js
@@ -3,21 +3,24 @@
  * - some important command features may be changed/broken in the future
  */
 
+/*
+ * Check out the {@link wiki https://discordjs.guide/} for information on `require('discord.js)`
+ */
+const { ApplicationCommandType, ApplicationCommandOptionType, EmbedBuilder } = require('discord.js');
 
 /*
 Access global variables; By default, these are asyncLogs, config, & prefix
-  - `asyncLogs` - Replaces console to add more functionality
-    - startWatch(name[string], time[process.uptime], callback[function])
-    - stopWatch(name[string], time[process.uptime], callback[function], log[bool])
-  - `config`    - `config.json` is often called, this removes the need for require
-  - `prefix`    - the prefix of the bot [This will change in the future]
+    - `asyncLogs` - Replaces console to add more functionality
+        - startWatch(name[string], time[process.uptime], callback[function])
+        - stopWatch(name[string], time[process.uptime], callback[function], log[bool])
+    - `config` - `config.json` is often called, this removes the need for require
+    - `prefix` - the prefix of the bot [This will change in the future]
 */
 const { log, debug, warn, error, startWatch, stopWatch } = asyncLogs;
 
 const { statusMessages, github, version } = config;
 
 // if (prefix === "!") log("\tExample: The prefix is !"); // This runs on command load
-
 
 /**
  * This handles the command
@@ -38,9 +41,9 @@ const run = function example (client, message, args) { // "run" is required (whe
  * 
  * @example
  * const conf = {
- *    enabled: true,  // enable & disable the command
- *    permLevel: 1,   // To-Be Implemented (No Use)
- *    type: 5         // To-Be Implemented (No Use)
+ *     enabled: true,    // enable & disable the command
+ *     permLevel: 1,     // To-Be Implemented (No Use)
+ *     type: 5           // To-Be Implemented (No Use)
  * }
  */
 const conf = { // The only required conf config is "enabled" (when true)
@@ -58,13 +61,13 @@ const conf = { // The only required conf config is "enabled" (when true)
  * const help = {                                   // The prefix is added in the help command
  *     name: `example`,                             // No spaces or lowercase, the prefix is added automatically
  *     aliases: [                                   // Make sure aliases are lower-case & have no spaces
- *          "othername",
- *          "anothername"
+ *         "othername",
+ *         "anothername"
  *     ],
  *     description: `Get the list of commands.`,    // The description shown using the !help command
- *     usage: `help [command]`                      // <> for mandatory & [] for optional
+ *     usage: `example [option]`                    // <> for mandatory & [] for optional
  * 
- *  // `omit` is a temporary solution
+ *     // `omit` is a temporary solution
  *     omit: true                                   // Hides the command from !help
  * }
  */
@@ -73,4 +76,22 @@ const help = { // The only required help config is "name" (when enabled)
     description: `This is an example command.`
 }
 
-module.exports = {run, conf, help} // this is a clean way to export multiple at once
+// -------------------------------------------------------------------------
+
+/**
+ * - This is where the slash [/] command settings go.
+ * - You do not need to change the script name to change the command name, just change `command.name`
+ */
+const command = {
+    name: 'example',
+    type: ApplicationCommandType.ChatInput,
+    options: [
+        {
+            name: 'option',
+            description: 'A sample option.',
+            type: ApplicationCommandOptionType.String,
+        }
+    ],
+};
+
+module.exports = {run, conf, help, command} // this is a clean way to export multiple at once