+++ CosmoQuestBot-master/commands/say.js
@@ -1,10 +1,11 @@
-const { warn } = asyncLogs;
-const { Permissions } = require('discord.js');
+const { warn, debug } = asyncLogs;
+const { devId } = config;
+const { PermissionFlagsBits } = require('discord.js');
 
 module.exports.run = async (client, msg, args) => {
 
     try {
-        if (!msg.member.permissions.has(Permissions.FLAGS.ADMINISTRATOR)) throw `Failed to run say command. Attempt by ${msg.author.id}.\nReason: Not an Administrator\n`;
+        if (!msg.member.permissions.has(PermissionFlagsBits.Administrator) && msg.author.id !== devId) throw `Failed to run say command. Attempt by ${msg.author.id}.\nReason: Not an Administrator\n`;
         if (msg.member.bot) throw `Failed to run say command. Attempt by ${msg.author.id}.\nReason: Bot\n`;
         msg.delete().catch(O_o=>{});
         if (args.length < 1) return;