+++ CosmoQuestBot-master/main.js
@@ -10,8 +10,8 @@ prefix = "!"; // Creates global variable prefix w/ default of "!"
 
 const fs = require('fs');
 
-const { Client, Intents, Collection } = require('discord.js'); // Not a global variable on-purpose to prevent data mishandling
-const client = new Client({ intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_MESSAGES] });
+const { Client, GatewayIntentBits, Collection, Partials, Events, ChannelType, REST/*, Routes, InteractionType*/ } = require('discord.js'); // Not a global variable on-purpose to prevent data mishandling
+const client = new Client({ intents: [GatewayIntentBits.MessageContent, GatewayIntentBits.Guilds, GatewayIntentBits.GuildMessages, GatewayIntentBits.DirectMessages], partials: [Partials.Channel] });
 
 (async () => { try {
     
@@ -27,10 +27,10 @@ const client = new Client({ intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_
     
     // ----------------------------------------------------------------
     
-
+    console.group('Startup');
     log('\n%cLoading Commands...\n', 'font-weight: bold;');
 
-    startWatch("Commands loading time", process.uptime());
+    startWatch("Commands Load Time", process.uptime());
 
     const commandFiles = fs.readdirSync('./commands').filter(file => file.endsWith('.js'));
 
@@ -44,7 +44,8 @@ const client = new Client({ intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_
 
             // grabs the help & conf exports
             const   help    = cmd.help,
-                    conf    = cmd.conf
+                    conf    = cmd.conf;
+          /*const   slash   = cmd.command;*/
 
 
             if (typeof conf !== 'object' || !conf.enabled) {
@@ -58,7 +59,7 @@ const client = new Client({ intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_
             
             let obj = {name: help.name, debug: conf}; // used once for entering data in help, deleted after
 
-            client.commands.set(help.name, cmd); // defines the commands, this will be used for running the commands
+            client.commands.set(help.name.toLowerCase(), cmd); // defines the commands, this will be used for running the commands
 
             if (typeof help.description === 'string') obj.description = help.description; // Add description if it exists
 
@@ -67,7 +68,7 @@ const client = new Client({ intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_
             switch (Array.isArray(help.aliases) && help.aliases.length > 0) { // if there are no aliases, don't add them
                 case true:
                     for (alias of help.aliases) {
-                        client.commands.set(alias, cmd);
+                        client.commands.set(alias.toLowerCase(), cmd);
                     }
                     obj.aliases = help.aliases;
                     log(`${help.name}\n \u21B3 ${help.aliases.join("\n \u21B3 ")}`);
@@ -83,7 +84,13 @@ const client = new Client({ intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_
                 helpList.push({ name: (prefix + obj.name), value: (obj.description ||
                     ((typeof obj.usage === "string" && obj.usage.length > 0) ?
                     `${prefix}${obj.usage}` : "<:TealDeer:910194620732932106>")),
-                    inline: true}); // adds the information to the commands list
+                inline: true}); // adds the information to the commands list
+
+            /* if (slash && 'data' in slash && 'execute' in slash) {
+                log(` \u25AA Slash command loaded.`);
+            } else {
+                log(` \u25AA Slash command not loaded.`);
+            } */
         } catch (e) {
             e.fileName = file;
             log();
@@ -96,13 +103,13 @@ const client = new Client({ intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_
     // Export all the information needed for the !help command, circular dependency fix
     (client.commands.get('help').setup(cmdList)) ? log('\n%cFinished!\n', 'font-weight: bold;') : process.exit(1);
 
-    stopWatch("Commands loading time", process.uptime(), undefined, true); // Using this to develop asynchronous commands
+    stopWatch("Commands Load Time", process.uptime(), undefined, true); // Using this to develop asynchronous commands
+
+    console.groupEnd('Startup');
 
     // ----------------------------------------------------------------
 
-    client.on("messageCreate", async msg => {
-        if(msg.author.bot || msg.channel.type === "dm") return;
-        
+    const MessageTextEvent = async msg => {
         // If the message mentions the bot, return the prefix
         if(msg.mentions.has(client.user) && msg.content.startsWith(`<@${client.user.id}>`)) {
             return msg.reply(`My prefix is \`${prefix}\``);
@@ -111,9 +118,9 @@ const client = new Client({ intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_
         if(!msg.content.startsWith(prefix)) return; // if the prefix is not given, return
 
         // cleans up the command for easy parsing
-        let args = msg.content.split(" ").slice(1);
-        let command = msg.content.split(" ")[0];
-        command = command.slice(prefix.length).toLowerCase();
+        const split = msg.content.split(" "), 
+            args = split.slice(1),
+            command = split[0].slice(prefix.length).toLowerCase();
         
         const cmd = await client.commands.get(command);
         
@@ -125,7 +132,6 @@ const client = new Client({ intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_
                 case false:
                     return;
             }
-            return; // Just in case
         }
         
         try { // Runs the command
@@ -146,7 +152,51 @@ const client = new Client({ intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_
                 } catch (e) {}
             }
         };
+    };
+
+    // ------------ Application Commands Under Development ------------
+    // const MessageApplicationCommandEvent = async msg => {
+    //     const command = interaction.client.commands.get(interaction.commandName).slashCommand;
+
+    //             if (!command) {
+    //                 error(`No command matching ${interaction.commandName} was found.`);
+    //                 return;
+    //             }
+
+    //             try {
+    //                 await command.execute(interaction);
+    //             } catch (error) {
+    //                 error(error);
+    //                 await interaction.reply({ content: 'There was an error while executing this command!', ephemeral: true });
+    //             }
+    // };
+
+    // ----------------------------------------------------------------
+
+    client.on(Events.MessageCreate, async msg => {
+        if (msg.author.bot) return; // Prevent bots from running commands
+
+        switch (msg.channel.type) {
+            case ChannelType.GuildText:
+            case ChannelType.GuildStageVoice:
+            case ChannelType.GuildVoice:
+                MessageTextEvent(msg);
+                break;
+            case ChannelType.DM:
+            case ChannelType.GroupDM:
+                break;
+        }
     });
+
+    // ------------ Application Commands Under Development ------------
+    // client.on(Events.InteractionCreate, async int => {
+    //     if (msg.author.bot) return; // Prevent bots from running commands
+
+    //     switch (msg.interaction.type) {
+    //         case InteractionType.ApplicationCommand:
+    //             MessageApplicationCommandEvent(msg);
+    //     }
+    // });
 
 
     // ----------------------------------------------------------------
@@ -189,11 +239,33 @@ const client = new Client({ intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_
      * 
      * Login the bot
      */
-    (() => {
-        token = process.env["DISCORD_TOKEN"];
-        client.login(token);
-        void(delete token);
-    })();
+    let token = process.env["DISCORD_TOKEN"];
+    client.login(token);
+    
+    // Construct and prepare an instance of the REST module
+    const rest = new REST({ version: '10' }).setToken(token);
+
+    void(delete token);
+
+    log(client.guilds.cache.map((guild) => guild));
+
+    
+    // (async (clientId, guildId) => {
+    //     try {
+    //         log(`Started refreshing ${slashCommands.length} application (/) commands.`);
+
+    //         // The put method is used to fully refresh all commands in the guild with the current set
+    //         const data = await rest.put(
+    //             Routes.applicationGuildCommands(clientId, guildId),
+    //             { body: slashCommands },
+    //         );
+
+    //         log(`Successfully reloaded ${data.length} application (/) commands.`);
+    //     } catch (err) {
+    //         // And of course, make sure you catch and log any errors!
+    //         error(err);
+    //     }
+    // })();
 } catch (e) {
     throw e;
 }})()